+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Compatibility:
 * Removed extra public methods on `String`.
 * Implemented `rb_array_sort` and `rb_array_sort_bang`.
 * Do not create a finalizers `Thread` if there are other public languages, which is helpful for polyglot cases (#2035).
+* Implemented `rb_enc_isalnum` and `rb_enc_isspace`.
 
 Performance: